backend/requirements.txt
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Login==0.6.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
zipp==3.17.0
